% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_resource}
\alias{apigateway_delete_resource}
\title{Deletes a Resource resource}
\usage{
apigateway_delete_resource(restApiId, resourceId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] [Required] The identifier of the Resource resource.}
}
\description{
Deletes a Resource resource.
}
\section{Request syntax}{
\preformatted{apigateway$delete_resource(
  restApiId = "string",
  resourceId = "string"
)
}
}

\keyword{internal}
