% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_rest_api}
\alias{apigateway_delete_rest_api}
\title{Deletes the specified API}
\usage{
apigateway_delete_rest_api(restApiId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}
}
\description{
Deletes the specified API.
}
\section{Request syntax}{
\preformatted{apigateway$delete_rest_api(
  restApiId = "string"
)
}
}

\keyword{internal}
