% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_account}
\alias{apigateway_get_account}
\title{Gets information about the current Account resource}
\usage{
apigateway_get_account()
}
\description{
Gets information about the current Account resource.
}
\section{Request syntax}{
\preformatted{apigateway$get_account()
}
}

\keyword{internal}
