% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_method}
\alias{apigateway_get_method}
\title{Describe an existing Method resource}
\usage{
apigateway_get_method(restApiId, resourceId, httpMethod)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] [Required] The Resource identifier for the Method resource.}

\item{httpMethod}{[required] [Required] Specifies the method request's HTTP method type.}
}
\description{
Describe an existing Method resource.
}
\section{Request syntax}{
\preformatted{apigateway$get_method(
  restApiId = "string",
  resourceId = "string",
  httpMethod = "string"
)
}
}

\keyword{internal}
