% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_usage_plan}
\alias{apigateway_get_usage_plan}
\title{Gets a usage plan of a given plan identifier}
\usage{
apigateway_get_usage_plan(usagePlanId)
}
\arguments{
\item{usagePlanId}{[required] [Required] The identifier of the UsagePlan resource to be retrieved.}
}
\description{
Gets a usage plan of a given plan identifier.
}
\section{Request syntax}{
\preformatted{apigateway$get_usage_plan(
  usagePlanId = "string"
)
}
}

\keyword{internal}
