% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_vpc_link}
\alias{apigateway_update_vpc_link}
\title{Updates an existing VpcLink of a specified identifier}
\usage{
apigateway_update_vpc_link(vpcLinkId, patchOperations)
}
\arguments{
\item{vpcLinkId}{[required] [Required] The identifier of the VpcLink. It is used in an Integration
to reference this VpcLink.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\description{
Updates an existing VpcLink of a specified identifier.
}
\section{Request syntax}{
\preformatted{apigateway$update_vpc_link(
  vpcLinkId = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

\keyword{internal}
